Changed socket used for clock to game namespace
