+++ game/src/server/namespace/RoomSelectionNamespace.ts
@@ -1,8 +1,5 @@
 import { Server as SocketIOServer, Socket } from "socket.io";
 import RoomRepository from "../data/RoomRepository";
-import { TimeRequestEvent, TimeResponseEvent } from "./../../communication/clock/DataInterfaces";
-import { CLIENT_EVENT_NAMES as CE, SERVER_EVENT_NAMES as SE } from "./../../communication/clock/EventNames";
-import { Clock } from "./../../gameCore/clock/Clock";
 
 export default class RoomSelectionNamespace {
 	private io: SocketIOServer;
@@ -24,15 +21,6 @@ export default class RoomSelectionNamespace {
 				"room-update",
 				this.roomRepo.getAllRooms().map((room) => `Id: ${room.getRoomId()} - Currently in: ${room.getGameState()}`)
 			);
-
-			socket.on(SE.TIME_REQUEST, (timeRequestEvent: TimeRequestEvent) => {
-				const serverCurrentLocalTime = Clock.now();
-				socket.emit(CE.TIME_RESPONSE, <TimeResponseEvent>{
-					clientCurrentLocalTime: timeRequestEvent.clientCurrentLocalTime,
-					serverCurrentLocalTime: serverCurrentLocalTime,
-				});
-				console.log(`[TIME_RESPONSE] Client:${timeRequestEvent.clientCurrentLocalTime}, Server:${serverCurrentLocalTime}`);
-			});
 
 			socket.on("disconnect", () => {
 				console.log("room selection disconnection");
