+++ game/src/server/rooms/RoomManager.ts
@@ -1,6 +1,9 @@
 import { Server as SocketIOServer, Socket } from "socket.io";
+import { TimeRequestEvent, TimeResponseEvent } from "../../communication/clock/DataInterfaces";
 import UserInfo from "../../communication/userInfo";
+import { Clock } from "../../gameCore/clock/Clock";
 import RoomRepository from "../data/RoomRepository";
+import { CLIENT_EVENT_NAMES as CE, SERVER_EVENT_NAMES as SE } from "./../../communication/clock/EventNames";
 import RoomFactory from "./RoomFactory";
 
 export default class RoomManager {
@@ -29,6 +32,15 @@ export default class RoomManager {
 
 			socket.on("disconnect", () => {
 				console.log("disconnected");
+			});
+
+			socket.on(SE.TIME_REQUEST, (timeRequestEvent: TimeRequestEvent) => {
+				const serverCurrentLocalTime = Clock.now();
+				socket.emit(CE.TIME_RESPONSE, <TimeResponseEvent>{
+					clientCurrentLocalTime: timeRequestEvent.clientCurrentLocalTime,
+					serverCurrentLocalTime: serverCurrentLocalTime,
+				});
+				console.log(`[TIME_RESPONSE(RoomManager)] Client:${timeRequestEvent.clientCurrentLocalTime}, Server:${serverCurrentLocalTime}`);
 			});
 
 			socket.on("create-room", () => {
